Refactor: leading_trivia and trailing_trivia return all trivia pieces
This is opposed to the previous implementation where the leading trivia would
only return the immediate leading trivia stored in the token, which could
be confusing. To get the trivia between two tokens, the user would have to
manually extract leading and trailing trivias for the current and the
previous resp. next token.